Allocate only one resource manager.
Originally, one was allocated by run(), because we needed its messages
attribute to be reset each time run() was called.  However, by changing
messages to a property which clears the messages list each time it is
accessed, we eliminated that problem.

This commit also adds the forgotten call to release_all() to clean up
all resources.